requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
geoalchemy2==0.15.2

# AWS
boto3==1.35.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Utilities
httpx==0.27.2
Pillow==10.4.0
python-magic==0.4.27

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0

# Linting
ruff==0.6.9
